pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = [".venv*", "tests*"]

[project]
name = "my-personal-assistant"
version = "0.1.0"
description = "LLM 전문 AI 개발자를 위한 개인 비서"
requires-python = ">=3.13"
dependencies = [
    # 에이전트
    "deepagents>=0.1.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    # 백엔드
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    # 채팅 UI
    "chainlit>=2.0.0",
    # 검색
    "tavily-python>=0.5.0",
    # 논문 수집
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    # 크론잡
    "apscheduler>=3.10.0",
    # 유틸리티
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "fastmcp>=3.1.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",         # FastAPI 테스트용
]

[tool.ruff]
target-version = "py313"
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
exclude = ["frontend/app.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
    "types-pyyaml>=6.0.12.20250915",
]
